fix v0.1.49 compile — Resource.transferSize is Long not Int
The diagnostic field was Int but ResourceAdvertisement.transferSize
is Long. Trivial type widen.

+++ shared/src/commonMain/kotlin/io/github/thatsfguy/reticulum/engine/LinkSession.kt
@@ -84,7 +84,7 @@ class LinkSession internal constructor(
     private var lastRxAt: Long = -1L
     /** Set when CTX_RESOURCE_ADV arrived — tracks total parts + received parts. */
     private var resourceAdvParts: Int = -1
-    private var resourceAdvBytes: Int = -1
+    private var resourceAdvBytes: Long = -1L
     private var resourceChunksRx: Int = 0
 
     /**
